metodo ordena por peso

+++ src/BagagemDeMao.java
@@ -1,6 +1,6 @@
 public class BagagemDeMao extends Bagagem {
 
-    public BagagemDeMao(double peso, String descricao) {
+    public BagagemDeMao(int peso, String descricao) {
         super(peso, descricao);
     }
